Ignore trailing whitespace after a backslash and before EOL
A lot of formula files were passed around in the early days of the
internet when -- would you belive it? -- transmitting and sharing
of text files wasn't a guaranteed lossless operation.  Sometimes
things get dropped and sometimes EOL characters get messed wih.
A line ending in a continuing backslash (\) could result in extra
spaces being added when emailed, uploaded to a web site, copy
and pasted between application and so-on.