Assignment is a statement, not an expression
Unlike C where an assignment can be a valid expression and used as part
of a larger expression, the formula language does not allow this.
Assignment must be a statement, although it can be chained as z=c=pixel.
Assignment is right-associative, so the chained assignment is as if
c=pixel,z=c had been written.

Fixes #22