Allow right hand side of assignment to be a conjunctive expression
This is despite the documentation of = binding tighter than comparison
operators.  It turns out that the ad-hoc parser in FRACTINT didn't follow
the precedence more explicitly.

It's also moot since assignment can't be used as an expression, so after
you identify the right-hand side of an assignment statement you might
as well just evaluate any possible expression that can be written.